+++ pkgdown.yml
@@ -10,5 +10,5 @@ articles:
   performance: performance.html
   seecolorscales: seecolorscales.html
   seecolorscales_dark: seecolorscales_dark.html
-last_built: 2021-06-14T16:08Z
+last_built: 2021-06-15T18:28Z
 
