+++ R/plot.check_normality.R
@@ -18,8 +18,6 @@
 #' m <<- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 #' result <- check_normality(m)
 #' plot(result)
-#' @importFrom bayestestR estimate_density
-#' @importFrom stats residuals sd dnorm rstudent ppoints pnorm fitted
 #' @export
 plot.see_check_normality <- function(x,
                                      type = c("density", "qq", "pp"),
