+++ R/print.data_plot.R
@@ -1,5 +1,4 @@
 #' @method print data_plot
-#' @importFrom graphics plot
 #' @export
 print.data_plot <- function(x, ...) {
   print(as.data.frame(x))
