+++ R/plot.performance_pp_check.R
@@ -1,4 +1,3 @@
-#' @importFrom stats reshape
 #' @export
 data_plot.performance_pp_check <- function(x, ...) {
   columns <- colnames(x)
