### 2021-06-17 #3372 @danielskatz
Thanks @fboehm - can you go ahead and take it then, via the command ```@whedon assign me as editor```, and then remove the waitlisted tag?  (You can always do this with any paper that's not undergoing scope review (with the `query-scope` tag)
