### 2021-06-19 #3372 @fboehm
:wave: @trinker & @JohnCoene & @garretrc & @MatthewSmith430, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Thank you for considering this!
