### 2021-06-21 #3372 @fboehm
@cdcrabtree, @jakobbossek, @tweed1e - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Thank you for considering this!
