### 2021-06-22 #3393 @whedon
[REVIEW]: see: An R Package for Visualizing Statistical Models
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/easystats/see/" target ="_blank">https://github.com/easystats/see/</a>
**Version:** 0.6.5
