+++ pkgdown.yml
@@ -10,5 +10,5 @@ articles:
   performance: performance.html
   seecolorscales: seecolorscales.html
   seecolorscales_dark: seecolorscales_dark.html
-last_built: 2021-06-17T11:46Z
+last_built: 2021-06-23T03:06Z
 
