### 2021-06-29 3c6ab50
Fix one more `guides(<scale> = FALSE)`
