+++ R/print.check_model.R
@@ -271,7 +271,7 @@ print.see_check_model <- function(x,
         detrend = detrend
       )
     )
-    y_lab <- "Sample - Normal Distribution Quantiles"
+    y_lab <- "Sample Quantiles"
   } else {
     message(
       "For confidence bands",
