+++ R/theme_blackboard.R
@@ -18,6 +18,7 @@ theme_blackboard <-
            plot.title.size = 15,
            plot.title.face = "plain",
            plot.title.space = 20,
+           plot.title.position = "plot",
            legend.position = "right",
            axis.title.space = 20,
            legend.title.size = 13,
@@ -34,6 +35,7 @@ theme_blackboard <-
       plot.title.size = plot.title.size,
       plot.title.face = plot.title.face,
       plot.title.space = plot.title.space,
+      plot.title.position = plot.title.position,
       legend.position = legend.position,
       axis.title.space = axis.title.space,
       legend.title.size = legend.title.size,
