+++ R/theme_lucid.R
@@ -17,6 +17,7 @@ theme_lucid <- function(base_size = 11,
                         plot.title.size = 12,
                         plot.title.face = "plain",
                         plot.title.space = 15,
+                        plot.title.position = "plot",
                         legend.position = "right",
                         axis.title.space = 10,
                         legend.title.size = 11,
@@ -33,6 +34,7 @@ theme_lucid <- function(base_size = 11,
     plot.title.size = plot.title.size,
     plot.title.face = plot.title.face,
     plot.title.space = plot.title.space,
+    plot.title.position = plot.title.position,
     legend.position = legend.position,
     axis.title.space = axis.title.space,
     legend.title.size = legend.title.size,
