### 2021-06-29 7aee60e
Add plot.check_model method
