+++ R/plot.check_model.R
@@ -1,0 +1,499 @@
+#' @importFrom rlang .data
+#' @export
+plot.see_check_model <- function(x,
+                                  style = theme_lucid,
+                                  colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
+                                  ...) {
+  orig_x <- x
+  p <- list()
+
+  panel <- attr(x, "panel")
+  check <- attr(x, "check")
+  size_point <- attr(x, "dot_size")
+  size_line <- attr(x, "line_size")
+  size_text <- attr(x, "text_size")
+  alpha_level <- attr(x, "alpha")
+  dot_alpha_level <- attr(x, "dot_alpha")
+  detrend <- attr(x, "detrend")
+
+  if (missing(style) && !is.null(attr(x, "theme"))) {
+    theme_style <- unlist(strsplit(attr(x, "theme"), "::", fixed = TRUE))
+    style <- get(theme_style[2], asNamespace(theme_style[1]))
+  }
+
+  if (missing(colors)) {
+    colors <- attr(x, "colors")
+  }
+
+  if (is.null(colors)) {
+    colors <- c("#3aaf85", "#1b6ca8", "#cd201f")
+  }
+
+  colors <- unname(colors)
+
+  if (is.null(alpha_level)) {
+    alpha_level <- .2
+  }
+
+  if (is.null(dot_alpha_level)) {
+    dot_alpha_level <- .8
+  }
+
+  if (is.null(check)) check <- "all"
+
+  if ("NCV" %in% names(x) && any(c("ncv", "linearity", "all") %in% check)) {
+    p$NCV <- .plot_diag_linearity(
+      x$NCV,
+      size_point,
+      size_line,
+      alpha_level,
+      theme_style = style,
+      colors = colors,
+      dot_alpha_level = dot_alpha_level
+    )
+  }
+
+  if ("HOMOGENEITY" %in% names(x) && any(c("homogeneity", "all") %in% check)) {
+    p$HOMOGENEITY <- .plot_diag_homogeneity(
+      x$HOMOGENEITY,
+      size_point,
+      size_line,
+      alpha_level,
+      theme_style = style,
+      colors = colors,
+      dot_alpha_level = dot_alpha_level
+    )
+  }
+
+  if ("VIF" %in% names(x) && any(c("vif", "all") %in% check)) {
+    p$VIF <- .plot_diag_vif(
+      x$VIF,
+      theme_style = style,
+      colors = colors
+    )
+  }
+
+  if ("OUTLIERS" %in% names(x) && any(c("outliers", "all") %in% check)) {
+    p$OUTLIERS <- .plot_diag_outliers_new(
+      x$INFLUENTIAL,
+      size_text = size_text,
+      size_line = size_line,
+      theme_style = style,
+      colors = colors,
+      dot_alpha_level = dot_alpha_level
+    )
+  }
+
+  if ("QQ" %in% names(x) && any(c("qq", "all") %in% check)) {
+    p$QQ <- .plot_diag_qq(
+      x$QQ,
+      size_point,
+      size_line,
+      alpha_level = alpha_level,
+      detrend = detrend,
+      theme_style = style,
+      colors = colors,
+      dot_alpha_level = dot_alpha_level
+    )
+  }
+
+  if ("NORM" %in% names(x) && any(c("normality", "all") %in% check)) {
+    p$NORM <- .plot_diag_norm(
+      x$NORM,
+      size_line,
+      alpha_level = alpha_level,
+      theme_style = style,
+      colors = colors
+    )
+  }
+
+  if ("REQQ" %in% names(x) && any(c("reqq", "all") %in% check)) {
+    ps <- .plot_diag_reqq(
[…]
