### 2021-06-29 13cb9a0
Switch to patchwork instead of gridExtra
Closes https://github.com/easystats/see/issues/99
