+++ DESCRIPTION
@@ -70,7 +70,6 @@ Suggests:
     ggrepel,
     glmmTMB,
     grid,
-    gridExtra,
     httr,
     lavaan,
     lme4,
@@ -81,6 +80,7 @@ Suggests:
     modelbased,
     NbClust,
     nFactors,
+    patchwork,
     performance (>= 0.7.1),
     psych,
     qqplotr,
