+++ man/theme_modern.Rd
@@ -10,6 +10,7 @@ theme_modern(
   plot.title.size = 15,
   plot.title.face = "plain",
   plot.title.space = 20,
+  plot.title.position = "plot",
   legend.position = "right",
   axis.title.space = 20,
   legend.title.size = 13,
@@ -32,6 +33,8 @@ theme_modern(
 \item{plot.title.face}{Title font face ("plain", "italic", "bold", "bold.italic").}
 
 \item{plot.title.space}{Title spacing.}
+
+\item{plot.title.position}{Alignment of the plot title/subtitle and caption. The setting for plot.title.position applies to both the title and the subtitle. A value of "panel" (the default) means that titles and/or caption are aligned to the plot panels. A value of "plot" means that titles and/or caption are aligned to the entire plot (minus any space for margins and plot tag).}
 
 \item{legend.position}{the position of legends ("none", "left", "right",
 "bottom", "top", or two-element numeric vector)}
