### 2021-06-29 ae49d37
Remove gridExtra from check_distribution
