### 2021-06-29 ffde749
use geom_histogram instead of geom_bar
Works much better for non-discrete scales.
