+++ R/plot.check_model.R
@@ -4,7 +4,6 @@ plot.see_check_model <- function(x,
                                   style = theme_lucid,
                                   colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                   ...) {
-  orig_x <- x
   p <- list()
 
   panel <- attr(x, "panel")
