### 2021-06-29 7fa99c7
update tests for the new title position in plots
