### 2021-07-01 b313a9c
Fix regression from ffc433c94d54a4f2d7a3193c21a5da2c385bb73b
Closes https://github.com/easystats/see/issues/163
