+++ R/print.check_model.R
@@ -3,6 +3,6 @@ print.see_check_model <- function(x,
                                   style = theme_lucid,
                                   colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                   ...) {
-  plot(x, style = style, colors = colors, ...)
+  suppressWarnings(suppressMessages(print(plot(x, style = style, colors = colors, ...))))
   invisible(x)
 }
