+++ tests/testthat/test-plot.n_factors.R
@@ -2,7 +2,17 @@ test_that("`plot.see_n_factors()` works", {
   if (require("parameters") && require("nFactors")) {
     data(mtcars)
     result <- n_factors(mtcars, type = "PCA")
-    result
-    expect_s3_class(plot(result, type = "line"), "gg")
+
+    set.seed(123)
+    vdiffr::expect_doppelganger(
+      title = "bar graph",
+      fig = plot(result)
+    )
+
+    set.seed(123)
+    vdiffr::expect_doppelganger(
+      title = "line graph",
+      fig = plot(result, type = "line")
+    )
   }
 })
