### 2021-07-09 #3393 @MatthewSmith430
I would firstly like to commend the package contributors on their efforts to create a package which provides a clear and straightforward workflow for the visualisation of model parameters, predictions, and performance diagnostics. I applied the see functions to my own data using other models (not presented in the examples), such as ergms and found that the see package produced aesthetically pleasing plots.

@strengejacke please see below for my review comments, chiefly to help enhance the paper and documentation.

It would be very useful to clarify whether it is possible to edit the parameter names in the plots; this is related to the examples provided in the “Plotting Functions for the ‘parameters’ Package” and “Plotting Functions for the ‘correlation’ Package” vignettes.

When running the commands for model based, it prints the following message, rather than the plots shown in the readme/software paper: "Plots are coming in the future version." These details should be provided in the readme.

Software paper:
- In the summary I think there should be an opening sentence noting that see is embedded within the easystat universe/collection of packages (prior to the description of easystat).
- In the Statement of Need section of the paper you could also explicitly refer to the fortify function, which reflected a need for ggplot capabilities for model output, as a function exists to transform model output.

Readme:
- A very minor point – in the installation instructions, you could include a line that See is available on CRAN and can be installed simply, using install.packages('see'). For the development version available on GitHub, use the devtools package for installation.
- You may wish to note in the installation instruction that whilst it does not depend on easystat packages, it imports them.
- In the modelbased section of the readme it appears that the emmeans package is required for the examples; this needs to be highlighted.
- In the radar chart example, it would be useful to adjust this so that the labels are all completely visible. Sepal.Width is partially covered.

Documentation:
- It would be useful to see some additional text, especially introductory text on the vignette articles, even if repeated some of the details that were already in the readme.
