+++ NAMESPACE
@@ -76,6 +76,7 @@ export(bluebrown_colors)
 export(coord_radar)
 export(data_plot)
 export(flat_colors)
+export(geom_binomdensity)
 export(geom_count2)
 export(geom_count_borderless)
 export(geom_from_list)
@@ -175,3 +176,6 @@ export(theme_radar)
 export(theme_radar_dark)
 import(ggplot2)
 importFrom(rlang,.data)
+importFrom(stats,density)
+importFrom(stats,na.omit)
+importFrom(stats,xtabs)
