### 2021-07-11 ba7974d
address README issues in JOSS review
