+++ README.Rmd
@@ -58,31 +58,33 @@ The aim of the *see* package is to produce visualizations for a wide variety of 
 
 The central goal of *easystats* is to make the task of doing statistics in R as easy as possible. This goal is realized through intuitive and consistent syntax, consistent and transparent argument names, comprehensive documentation, informative warnings and error messages, and smart functions with sensible default parameter values. The *see* package follows this philosophy by using a single access point---the generic `plot()` method---for visualization of all manner of statistical results supported by *easystats*.
 
-## Installation
+# Installation
 
 [![CRAN](http://www.r-pkg.org/badges/version/see)](https://cran.r-project.org/package=see)
 ![R-check](https://github.com/easystats/see/workflows/R-check/badge.svg)
 
-Run the following:
+The *see* package is available on CRAN, while its development version is available on GitHub. In order to download the package from GitHub, you will need to have installed the *remotes* package first (`install.packages("remotes")`).
 
-```{r eval=FALSE}
-install.packages("devtools")
-devtools::install_github("easystats/see")
-```
+Type | Source | Command
+---|---|---
+Release | CRAN | `install.packages("datawizard")`
+Development | GitHub | `remotes::install_github("easystats/datawizard")`
 
-And then load it:
+Once you have downloaded the package, you can then load it using:
 
-```{r message=FALSE, warning=FALSE, eval=TRUE}
+```{r, eval=TRUE}
 library("see")
 ```
 
+This package provides visualisation toolbox for R packages in the `easystats` [ecosystem](https://easystats.github.io/easystats/). But note that it will not download any of the *easystats* packages for you. Rather, it will be loaded when a respective plotting method is requested by an *easystats* package.
+
 # Documentation
 
-This package provides visualisation toolbox for R packages in the `easystats` ecosystem and extra `geom`s, themes and color palettes for `ggplot2`
+Following resources are a good way to get to know more about the functionality offered by this package:
 
 [![Documentation](https://img.shields.io/badge/documentation-see-orange.svg?colorB=E91E63)](https://easystats.github.io/see/)
-[![Blog](https://img.shields.io/badge/blog-easystats-orange.svg?colorB=FF9800)](https://easystats.github.io/blog/posts/)
 [![Features](https://img.shields.io/badge/features-see-orange.svg?colorB=2196F3)](https://easystats.github.io/see/reference/index.html)
+[![Blog](https://img.shields.io/badge/blog-easystats-orange.svg?colorB=FF9800)](https://easystats.github.io/blog/posts/)
 
 # Plotting functions for 'easystats' packages
 
@@ -171,13 +173,15 @@ Plotting functions for the **effectsize** package are demonstrated [in this vign
 
 ## [modelbased](https://github.com/easystats/modelbased)
 
-The *modelbased* package computes model-based estimates and predictions from fitted models [@Makowski2020modelbased]. *see* provides methods to quickly visualize these model predictions.
+The *modelbased* package computes model-based estimates and predictions from fitted models [@Makowski2020modelbased]. *see* provides methods to quickly visualize these model predictions. For the following example to work, you need to have installed the *emmeans* package first.
 
 ```{r modelbased1}
 library(modelbased)
 library(see)
 
 model <- lm(mpg ~ wt * as.factor(cyl), data = mtcars)
+
+# for this to work, *emmeans* package should be available
 predicted <- estimate_prediction(model)
 
 plot(predicted)
@@ -195,7 +199,7 @@ Plotting functions for the **modelbased** package are demonstrated [in this vign
 
 ## [correlation](https://github.com/easystats/correlation)
 
[…]
