### 2021-07-11 cac101b
minor code reorg
