### 2021-07-11 aabd61a
address paper related feedback; need to reknit later
