+++ R/geom_binomdensity.R
@@ -2,7 +2,7 @@
 #'
 #' @param data A dataframe.
 #' @param x,y Characters corresponding to the x and y axis. Note that \code{y} must be a variable with two unique values.
-#' @param scale Method of scaling the dot-densities. Can be 'auto' (corresponding to the square root of the proportion), 'proportion', 'density' or a custom list with values for each facto level (see examples).
+#' @param scale Method of scaling the dot-densities. Can be 'auto' (corresponding to the square root of the proportion), 'proportion', 'density' or a custom list with values for each factor level (see examples).
 #' @param ... Other arguments passed to \code{ggdist::geom_dots}.
 #'
 #' @examples
