+++ vignettes/performance.Rmd
@@ -7,7 +7,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r, include = FALSE, echo=FALSE}
+```{r, include=FALSE, warning=FALSE, message=FALSE, echo=FALSE}
 knitr::opts_chunk$set(
   dpi = 300,
   out.width = "100%",
@@ -20,13 +20,23 @@ knitr::opts_chunk$set(
 )
 ```
 
+# Introduction
+
+A crucial aspect when building regression models is to evaluate the quality of modelfit. It is important to investigate how well models fit to the data and which fit indices to report. Functions to create diagnostic plots or to compute fit measures do exist, however, mostly spread over different packages. There is no unique and consistent approach to assess the model quality for different kind of models.
+
+The primary goal of the *performance* package in *easystats* ecosystem is to fill this gap and to provide utilities for computing **indices of model quality** and **goodness of fit**. These include measures like r-squared (R2), root mean squared error (RMSE) or intraclass correlation coefficient (ICC) , but also functions to check (mixed) models for overdispersion, zero-inflation, convergence or singularity.
+
+For more, see: <https://easystats.github.io/performance/>
+
+# Checking Model Assumptions
+
+Let's load the needed libraries first:
+
 ```{r setup}
 library(performance)
 library(lme4)
 library(see)
 ```
-
-# Checking Model Assumptions
 
 ## Binned Residuals
 
