+++ man/geom_binomdensity.Rd
@@ -9,9 +9,13 @@ geom_binomdensity(data, x, y, scale = "auto", ...)
 \arguments{
 \item{data}{A dataframe.}
 
-\item{x, y}{Characters corresponding to the x and y axis. Note that \code{y} must be a variable with two unique values.}
+\item{x, y}{Characters corresponding to the x and y axis. Note that \code{y}
+must be a variable with two unique values.}
 
-\item{scale}{Method of scaling the dot-densities. Can be 'auto' (corresponding to the square root of the proportion), 'proportion', 'density' or a custom list with values for each factor level (see examples).}
+\item{scale}{Method of scaling the dot-densities. Can be 'auto'
+(corresponding to the square root of the proportion), 'proportion',
+'density' or a custom list with values for each factor level (see
+examples).}
 
 \item{...}{Other arguments passed to \code{ggdist::geom_dots}.}
 }
@@ -26,10 +30,11 @@ data <- iris[1:100, ]
 
 ggplot() +
   geom_binomdensity(data,
-                    x = "Sepal.Length",
-                    y = "Species",
-                    fill = "red",
-                    color = NA)
+    x = "Sepal.Length",
+    y = "Species",
+    fill = "red",
+    color = NA
+  )
 
 # Different scales
 data[1:70, "Species"] <- "setosa" # Create unbalanced proportions
@@ -41,7 +46,8 @@ ggplot() +
 ggplot() +
   geom_binomdensity(data, x = "Sepal.Length", y = "Species", scale = "proportion")
 ggplot() +
-  geom_binomdensity(data, x = "Sepal.Length", y = "Species",
-                    scale = list("setosa" = 0.4, "versicolor" = 0.6))
-
+  geom_binomdensity(data,
+    x = "Sepal.Length", y = "Species",
+    scale = list("setosa" = 0.4, "versicolor" = 0.6)
+  )
 }
