### 2021-07-11 ba29e20
changing parameters names in correlation example
