### 2021-07-11 easystats/see#166 @bwiernik
> It would be very useful to clarify whether it is possible to edit the parameter names in the plots; this is related to the examples provided in

Perhaps let's add an example showing how these can be modified like in other ggplots?
