### 2021-07-11 easystats/see#166 @IndrajeetPatil
Done!

That led to https://github.com/easystats/see/issues/168
