+++ paper/paper.Rmd
@@ -133,7 +133,9 @@ model <- lm(wt ~ am * cyl, data = mtcars)
 # changing title and axis labels using ggplot2 functions
 plot(parameters(model)) +
   labs(title = "A Dot-and-Whisker Plot") + 
-  scale_x_discrete(labels = c("transmission * cylinders", "cylinders", "transmission"))
+  scale_x_discrete(labels = c("transmission * cylinders", 
+                              "cylinders", 
+                              "transmission"))
 ```
 
 \newpage
