+++ R/plot.compare_parameters.R
@@ -83,10 +83,9 @@ plot.see_compare_parameters <- function(x, show_intercept = FALSE, size_point = 
     x$Parameter <- factor(x$Parameter, levels = rev(unique(x$Parameter)))
   }
 
-  p <- ggplot(x, aes(x = .data$Parameter, y = .data$Coefficient, color = .data$group)) +
-    geom_hline(aes(yintercept = y_intercept), linetype = "dotted") +
-    geom_pointrange(aes(ymin = .data$CI_low, ymax = .data$CI_high), size = size_point, position = position_dodge(dodge_position)) +
-    coord_flip() +
+  p <- ggplot(x, aes(y = .data$Parameter, x = .data$Coefficient, color = .data$group)) +
+    geom_vline(aes(xintercept = y_intercept), linetype = "dotted") +
+    geom_pointrange(aes(xmin = .data$CI_low, xmax = .data$CI_high), size = size_point, position = position_dodge(dodge_position)) +
     theme_modern() +
     scale_color_material()
 
@@ -102,7 +101,7 @@ plot.see_compare_parameters <- function(x, show_intercept = FALSE, size_point = 
         colour = "black", hjust = "inward", size = size_text,
         position = position_dodge2(dodge_position)
       ) +
-      ylim(c(min(new_range), max(new_range)))
+      xlim(c(min(new_range), max(new_range)))
   }
 
   # check for exponentiated estimates. in such cases, we transform the y-axis
@@ -119,7 +118,8 @@ plot.see_compare_parameters <- function(x, show_intercept = FALSE, size_point = 
       new_range <- pretty(2 * max(x$CI_high))
       x_high <- which.max(max(new_range) < range)
     }
-    p <- p + scale_y_log10(
+    p <- p + scale_x_continuous(
+      trans = "log",
       breaks = range[x_low:x_high],
       limits = c(range[x_low], range[x_high]),
       labels = sprintf("%g", range[x_low:x_high])
@@ -130,7 +130,7 @@ plot.see_compare_parameters <- function(x, show_intercept = FALSE, size_point = 
   if (is.null(n_columns)) n_columns <- ifelse(sum(has_component, has_response, has_effects) > 1, 2, 1)
 
   if (ordinal_model) {
-    facet_scales <- "free_y"
+    facet_scales <- "free_x"
   } else {
     facet_scales <- "free"
   }
@@ -157,14 +157,14 @@ plot.see_compare_parameters <- function(x, show_intercept = FALSE, size_point = 
 
   if (isTRUE(axis_title_in_facet)) {
     p + labs(
-      x = "Parameter",
-      y = NULL,
+      y = "Parameter",
+      x = NULL,
       colour = "Model"
     )
   } else {
     p + labs(
-      x = "Parameter",
-      y = "Estimate",
+      y = "Parameter",
+      x = "Estimate",
       colour = "Model"
     )
   }
