### 2021-07-14 60496b2
fixes plot n factors method; closes #171
