### 2021-07-20 easystats/see#175 @jakobbossek
[JOSS review] Input checks
Unfortunately there are almost no sanity checks 😢  Passing inappropriate parameters hence may lead to error messages that are hard to understand.  Please use, e.g. [checkmate](https://github.com/mllg/checkmate) for fast and versatile argument checks.
