### 2021-07-20 easystats/see#175 @IndrajeetPatil
Hi, can you please give an example of what do you mean by there not being any sanity checks?
