### 2021-07-20 easystats/see#175 @jakobbossek
Sure.

```r
ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +  geom_violindot(scale = "arae")
```
throws the following error. The problem is a typo (arae vs. area), but the scale parameter is not checked for appropriate values. Hence, I get the following confusing error which does not give any hint what is actually wrong.
```
Fehler in `$<-.data.frame`(`*tmp*`, "xmaxv", value = numeric(0)) :
  Ersetzung hat 0 Zeilen, Daten haben 512
```

You can easily avoid this by adding the following to `geom_violindot`:
```r
checkmate::assert_choice(scale, choices = c("area", "some_other_valid_argument_value"))
```
