### 2021-07-21 acac435
closes #175; styling; removing old unused code
