+++ R/reshape_to_long.R
@@ -1,3 +1,5 @@
+# TODO: retire in favor of {datawizard}?
+
 #' @keywords internal
 .reshape_to_long <- function(x,
                              names_to = "group",
