+++ R/plot.rope.R
@@ -63,10 +63,6 @@ data_plot.rope <- function(x, data = NULL, show_intercept = FALSE, ...) {
 
 
 
-
-
-
-
 # Plot --------------------------------------------------------------------
 
 #' Plot method for Region of Practical Equivalence
