+++ R/geom_binomdensity.R
@@ -39,8 +39,14 @@
 #'   )
 #' @importFrom stats density na.omit xtabs
 #' @export
-geom_binomdensity <- function(data, x, y, scale = "auto", ...) {
+geom_binomdensity <- function(data,
+                              x,
+                              y,
+                              scale = c("auto", "density", "proportion"),
+                              ...) {
   insight::check_if_installed(c("ggplot2", "ggdist"))
+
+  scale <- match.arg(scale)
 
   # Sanitize y (e.g., if levels with no values, etc.)
   if (is.factor(data[[y]]) && length(levels(data[[y]])) > 2) {
@@ -82,8 +88,6 @@ geom_binomdensity <- function(data, x, y, scale = "auto", ...) {
 
 # Utilities ---------------------------------------------------------------
 
-
-
 .geom_binomdensity_scale <- function(data, x, y, scale = "auto") {
   prop <- prop.table(xtabs(paste("~", y), data)) # Get prop table (useful later)
   if (length(scale) == 1 && is.character(scale) && scale %in% c("density", "proportion", "auto")) {
@@ -95,6 +99,7 @@ geom_binomdensity <- function(data, x, y, scale = "auto", ...) {
       })
       prop <- prop / sum(prop)
     }
+
     # Square-rooted proportions
     if (scale == "auto") {
       prop <- sqrt(prop) / sum(sqrt(prop))
