+++ R/coord_radar.R
@@ -26,10 +26,16 @@ coord_radar <- function(theta = "x", start = 0, direction = 1, ...) {
   theta <- match.arg(theta, c("x", "y"))
   r <- ifelse(theta == "x", "y", "x")
 
-  ggplot2::ggproto("CordRadar", CoordPolar,
-    theta = theta, r = r, start = start,
+  ggplot2::ggproto(
+    "CordRadar",
+    CoordPolar,
+    theta = theta,
+    r = r,
+    start = start,
     direction = sign(direction),
-    is_linear = function(coord) TRUE,
+    is_linear = function(coord) {
+      TRUE
+    },
     ...
   )
 }
