+++ man/see_colors.Rd
@@ -13,7 +13,8 @@ see_colors(...)
 A character vector with color-codes.
 }
 \description{
-Can be used to get the hex code of specific colors from the See color palette. Use \code{see_colors()} to see all available color.
+Can be used to get the hex code of specific colors from the See color
+palette. Use \code{see_colors()} to see all available color.
 }
 \examples{
 see_colors()
