### 2021-07-21 easystats/see#178 @jakobbossek
[JOSS review] Improve documentation
The function documentation should be improved:

* Add links to referenced functions from other packages, e.g.instead of `\code{performance::check_homogeneity()}` write `\code{\link[performance]{performance::check_homogeneity}}`.
* Avoid plain URLs. Use links instead. E.g. in the description of `scale_color_material` replace `material design colors (https://www.materialui.co/color)` with `\href{https://www.materialui.co/color}{material design colors}`.
* Describe the expected data types. This is done sometimes, but not mostly not. E.g. in the roxygen2 docs for `plot.see_check_outliers` the expected type for `rescale_distance` is described (logical), but the type for `size_text` is missing (positive numeric). This is linked to #175 and would certainly improve user experience.
* Some parameter descriptions are not satisfactory. E.g. for `coord_radar` the parameter `theta` is described as follows: "Can be 'x' or 'y'." OK, but what does the parameter actually do?
