### 2021-07-21 #3393 @jakobbossek
Thanks for this package. I think it is of scientific interest for a good part of R users.

I agree with all minor points for improvement raised by @MatthewSmith430
Furthermore, I believe that the API/function documentation can be improved (see easystats/see#178) and in particular argument checks are missing throughout the package (see easystats/see#175). The latter is IMHO a problem in most R packages what makes debugging of R code really annoying.

My biggest concern at the moment is that the paper is almost identical to the README file (some parts are even copy & paste). IMHO that does not make a good impression! The authors should also check whether it is really necessary to place that many code examples and the produced figures in the paper. Again, this is done in the package vignettes and the long list of examples in the README. Instead I suggest to make the paper less crowded, but more focused. However, happy to discuss the latter point 😀👍
