### 2021-07-21 #3393 @IndrajeetPatil
Daniel can respond to this as well, but I just want to address the biggest concern you've raised here:

> My biggest concern at the moment is that the paper is almost identical to the README file (some parts are even copy & paste). IMHO that does not make a good impression!

This is **by choice**. Both `README` and the _JOSS_ paper are supposed to provide a quick overview about the package design and functionality, and we want to be as **consistent** across these two documents as possible. Because not everyone who is going to read the `README` document is also going to read the paper, and vice versa. In fact, our `README` document used to have different examples, we intentionally changed it while writing the paper to achieve the said consistency. I hope that rationale makes sense to you.
