### 2021-07-21 easystats/see#175 @jakobbossek
Actually, this is not fixed at all. The `scale` parameter was just an example. This commit did not change much.
Examples: in `coord_radar` the `start` parameter is not checked.
```r
if (require("dplyr") && require("tidyr") && require("ggplot2")) {
  data <- iris %>%
    group_by(Species) %>%
    summarise_all(mean) %>%
    pivot_longer(-Species)

  data %>%
    ggplot(aes(x = name, y = value, color = Species, group = Species)) +
    geom_polygon(fill = NA, size = 2) +
    coord_radar(start = "abc")
}
```

```
Error in x + coord$start :
  non numeric argument for binary operator
```

In `plot.see_easycorrelation I can pass negative values for `size_point`. This does not break the code, but it actually makes no sense.

There are many more examples.
