### 2021-07-21 easystats/see#175 @IndrajeetPatil
I am not sure if it makes sense to check nonsensical inputs to arguments because it is highly unlikely that anyone is actually going to make that mistake in the wild.

For example, `ggplot2` is used by millions of users and yet there is no check for `size` not being negative and for a good reason: it is highly unlikely that anyone is actually going to enter negative values in this context.

``` r
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point(size = -2)
```

![](https://i.imgur.com/L5395l5.png)

At least, that's my take on this issue. I will let other authors comment on this, and will keep the issue open for the same.
