### 2021-07-21 easystats/see#175 @jakobbossek
Sorry for being nitpicky, but this is what bothers me about the majority of R packages. Of course it is unlikely that some user calls, say `coord_radar` by passing a character (vector) for the `start` parameter. However, if the plot code is embedded in a loop or some larger script  and the author messed up something by passing down the wrong variable, a meaningful error message would help a lot to identify the issue quickly.

I really suggest to adopt something like [checkmate](https://github.com/mllg/checkmate). It is easy to use and very fast (see the [R Journal paper](https://journal.r-project.org/archive/2017/RJ-2017-028/index.html) for a benchmark). Alternatively just use `stopifnot`. E.g. for the `start` parameter it could be something like:
```r
stopifnot(is.numeric(start), (length(start) == 1))
```
