### 2021-07-21 easystats/see#175 @bwiernik
(I would really appreciate if you would use "input checks" rather than "sanity checks" as a term.)
