### 2021-07-21 easystats/see#175 @bwiernik
A few general thoughts:

1. We try to limit the number of dependencies in packages to the absolute minimum, so using checkmate isn't an option. Any checks would be implemented with tryCatch() or stopifnot()

2. For the geom_ and theme_ functions, these are really minor extensions to the broader ggplot2 framework. As a result, it is critical that these have consistent behavior with the underlying ggplot2 framework. Although a negative size value might not make much sense conceptually, allowing that is the behavior of every ggplot2 geom and aesthetic. It's not uncommon, for example, for users to use negative size values as a method to suppress rendering of some elements using negative size values. To prevent that would potentially produce jarring inconsistency with the behavior of other geoms.

   I encountered a similar issue recently with another ggplot2 extension. This package defined an unusual behavior for position_dodge(), which meant that its geoms did not align with base ggplot2 geoms when dodged. The package developer did this to ensure the new geoms never overflowed the plot area, but it made the package incompatible with the underlying ggplot2 framework. They needed to adjust the behavior so that dodge yielded consistent behavior and to create a new position yielded the previous "safer" behavior.

3. For input that actually yield errors, such as passing a character value to a coord_ `start` argument, compatibility is less of a concern. But I also think you are overstating the size of the concern. As stated above, the geom_, coord_, and theme_ functions in this package are minor extensions to underlying ggplot2 functions. A user that would encounter a problem with coord_radar() would encounter exactly the same issue with ggplot2::coord_polar(); a user that would encounter an issue with geom_point_borderless() would encounter exactly the same issue with ggplot2::geom_point(), etc. Adding type checks for these small ggplot2 functions (which I suspect will mostly be used internally by the model plotting functions that are the heart of this package) will not meaningfully affect user experience because they aren't present in any of the core ggplot2 functions. From my perspective, this would be an issue that would appropriately be fixed in ggplot2 itself at the stage of rendering a ggplot, not individually in every geom_, coord_, stat_, etc. call.

   Two additional comments here. One, importantly, evaluation of data in ggplot is not done until plot _rebdering_, not plot construction. In many common cases, such as a user passing arguments with `vars()` or an external vector, it's not possible to determine if input is valid because these arguments are passed lazily to ggproto().

   Two, as @IndrajeetPatil alluded to, ggplot2 is a very mature framework and has been in development for over 10 years. I think that if type mismatch issues like this were a major problem for users, it would have presented itself in that time. As it stands, I am very leery of implementing checks that go beyond recommended practices for writing ggplot2 extensions, especially if such checks would (1) be limited to a narrow range of allowable inputs (only bare objects, not names like `foo` or a `vars()` call), (2) would require potentially fragile checks to discern the check-able inputs from allowable-but-uncheckable inputs, and (3) might have negative performance consequences if we short-circuit the typical ggplot2 lazy evaluation workflow.
