### 2021-07-22 e1a90d4
update installation instruction
