### 2021-07-22 #3393 @strengejacke
Sorry for being so quiet here - I was on vacation, so my co-authors took the opportunity to address your raised issues. Let me briefly respond to your points, to summarize what we have done so far.

@MatthewSmith430 thank you for your review. Here are our replies:

_It would be very useful to clarify whether it is possible to edit the parameter names in the plots; this is related to the examples provided in the “Plotting Functions for the ‘parameters’ Package” and “Plotting Functions for the ‘correlation’ Package” vignettes._

We have added examples to both vignettes that show several options how to change the names of the parameters in the plot. Note that the example for the correlation vignette require the latest package version of the _correlation_ package, so these might not yet appear in the GitHub documentation page. We plan a new release cycle soon, so this issue is then resolved.

_When running the commands for model based, it prints the following message, rather than the plots shown in the readme/software paper: "Plots are coming in the future version." These details should be provided in the readme._

Thanks, this issue should also be resolved in the next release cycle, as plots have been implemented meanwhile.

_In the summary I think there should be an opening sentence noting that see is embedded within the easystat universe/collection of packages (prior to the description of easystat)._

We have revised the paragraph accordingly.

_In the Statement of Need section of the paper you could also explicitly refer to the fortify function, which reflected a need for ggplot capabilities for model output, as a function exists to transform model output._

We have also revised that paragraph accordingly.

_A very minor point – in the installation instructions, you could include a line that See is available on CRAN and can be installed simply, using install.packages('see'). For the development version available on GitHub, use the devtools package for installation.
You may wish to note in the installation instruction that whilst it does not depend on easystat packages, it imports them.
In the modelbased section of the readme it appears that the emmeans package is required for the examples; this needs to be highlighted._

We have revised the installation-section in the readme to clarify how to install the official CRAN release and the development version. Furthermore, we now use [`r-universe`](https://easystats.r-universe.dev/) for the installation of the development versions our packages, to simplify the installation process for novel users.

_In the radar chart example, it would be useful to adjust this so that the labels are all completely visible. Sepal.Width is partially covered._

We changed the example accordingly.

_It would be useful to see some additional text, especially introductory text on the vignette articles, even if repeated some of the details that were already in the readme._

Thanks for pointing this out, we revised the vignettes and added some more context about the specific packages and their purpose.

We hope we have addressed all raised issues to your satisfaction?
