### 2021-07-22 b3be0f0
remove plain URLs
