+++ man/palette_flat.Rd
@@ -7,14 +7,16 @@
 palette_flat(palette = "contrast", reverse = FALSE, ...)
 }
 \arguments{
-\item{palette}{Character name of palette. Depending on the color scale, can be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"}, \code{"contrast"} or \code{"light"} (for dark themes).}
+\item{palette}{Character name of palette. Depending on the color scale, can
+be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"},
+\code{"contrast"} or \code{"light"} (for dark themes).}
 
 \item{reverse}{Boolean indicating whether the palette should be reversed.}
 
 \item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
 }
 \description{
-The palette based on Flat UI colors (https://www.materialui.co/flatuicolors).
+The palette based on \href{https://www.materialui.co/flatuicolors}{Flat UI}.
 }
 \details{
 This function is usually not called directly, but from within
