+++ man/palette_metro.Rd
@@ -7,14 +7,17 @@
 palette_metro(palette = "complement", reverse = FALSE, ...)
 }
 \arguments{
-\item{palette}{Character name of palette. Depending on the color scale, can be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"}, \code{"contrast"} or \code{"light"} (for dark themes).}
+\item{palette}{Character name of palette. Depending on the color scale, can
+be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"},
+\code{"contrast"} or \code{"light"} (for dark themes).}
 
 \item{reverse}{Boolean indicating whether the palette should be reversed.}
 
 \item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
 }
 \description{
-The palette based on Metro colors (https://www.materialui.co/metrocolors).
+The palette based on \href{https://www.materialui.co/metrocolors}{Metro
+colors}.
 }
 \details{
 This function is usually not called directly, but from within
