+++ man/scale_color_see.Rd
@@ -36,7 +36,9 @@ scale_fill_see_d(palette = "contrast", discrete = TRUE, reverse = FALSE, ...)
 scale_fill_see_c(palette = "contrast", discrete = FALSE, reverse = FALSE, ...)
 }
 \arguments{
-\item{palette}{Character name of palette. Depending on the color scale, can be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"}, \code{"contrast"} or \code{"light"} (for dark themes).}
+\item{palette}{Character name of palette. Depending on the color scale, can
+be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"},
+\code{"contrast"} or \code{"light"} (for dark themes).}
 
 \item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}
 
