+++ README.md
@@ -98,10 +98,10 @@ results supported by *easystats*.
 The *see* package is available on CRAN, while its latest development
 version is available on R-universe (from *rOpenSci*).
 
-| Type        | Source   | Command                                                               |
-|-------------|----------|-----------------------------------------------------------------------|
-| Release     | CRAN     | `install.packages("see")`                                             |
-| Development | rOpenSci | `install.packages("see", repos = "https://easystats.r-universe.dev")` |
+| Type        | Source     | Command                                                               |
+|-------------|------------|-----------------------------------------------------------------------|
+| Release     | CRAN       | `install.packages("see")`                                             |
+| Development | R-universe | `install.packages("see", repos = "https://easystats.r-universe.dev")` |
 
 Once you have downloaded the package, you can then load it using:
 
