+++ README.Rmd
@@ -68,7 +68,7 @@ The *see* package is available on CRAN, while its latest development version is 
 Type | Source | Command
 ---|---|---
 Release | CRAN | `install.packages("see")`
-Development | rOpenSci | `install.packages("see", repos = "https://easystats.r-universe.dev")`
+Development | R-universe | `install.packages("see", repos = "https://easystats.r-universe.dev")`
 
 Once you have downloaded the package, you can then load it using:
 
