### 2021-07-22 0ac8631
use markdown syntax in roxygen docs
