+++ R/golden_ratio.R
@@ -2,7 +2,7 @@
 #'
 #' Returns the golden ratio (1.618034...). Useful to easily obtain golden
 #' proportions, for instance for a horizontal figure, if you want its height to
-#' be 8, you can set its width to be \code{golden_ratio(8)}.
+#' be 8, you can set its width to be `golden_ratio(8)`.
 #'
 #' @param x A number to be multiplied by the golden ratio. The default (x=1)
 #'   returns the value of the golden ratio.
