+++ R/plot.bayesfactor_models.R
@@ -1,23 +1,23 @@
 #' Plot method for Bayes Factors for model comparison
 #'
-#' The \code{plot()} method for the \code{bayestestR::bayesfactor_models()} function.
-#' These plots visualize the \strong{posterior probabilities} of the compared models.
+#' The `plot()` method for the `bayestestR::bayesfactor_models()` function.
+#' These plots visualize the **posterior probabilities** of the compared models.
 #'
 #' @param n_pies Number of pies.
 #' @param value What value to display.
 #' @param sort \describe{
-#'   \item{\emph{Plotting model parameters}}{
-#'   If \code{NULL}, coefficients are plotted in the order as they appear in the summary. Use \code{sort = "ascending"} (or \code{sort = TRUE})) resp. \code{sort = "descending"} to sort coefficients in ascending or descending order.
+#'   \item{*Plotting model parameters*}{
+#'   If `NULL`, coefficients are plotted in the order as they appear in the summary. Use `sort = "ascending"` (or `sort = TRUE`)) resp. `sort = "descending"` to sort coefficients in ascending or descending order.
 #'   }
-#'   \item{\emph{Plotting Bayes factors}}{
+#'   \item{*Plotting Bayes factors*}{
 #'   Sort pie-slices by posterior probability (descending)?
 #'   }
 #' }
 #' @param log Show log-transformed Bayes factors.
 #' @param prior_odds optional vector of prior odds for the models. See
-#'   \code{BayesFactor::priorOdds}. As the size of the pizza slices corresponds
+#'   `BayesFactor::priorOdds`. As the size of the pizza slices corresponds
 #'   to posterior probability (which is a function of prior probability and the BF),
-#'   custom \code{prior_odds} will change the slices' size.
+#'   custom `prior_odds` will change the slices' size.
 #' @inheritParams data_plot
 #'
 #' @return A ggplot2-object.

[…]
