+++ R/plot.easycormatrix.R
@@ -81,15 +81,14 @@ data_plot.see_easycormatrix <- function(x, data = NULL, digits = 3, size = 1, ..
 #' plot(s)
 #' @export
 plot.see_easycormatrix <- function(x,
-           show_values = FALSE,
-           show_p = FALSE,
-           show_legend = TRUE,
-           size_point = 1,
-           size_text = 3.5,
-           digits = 3,
-           type = c("circle", "tile"),
-           ...) {
-
+                                   show_values = FALSE,
+                                   show_p = FALSE,
+                                   show_legend = TRUE,
+                                   size_point = 1,
+                                   size_text = 3.5,
+                                   digits = 3,
+                                   type = c("circle", "tile"),
+                                   ...) {
   if (!"data_plot" %in% class(x)) {
     x <- data_plot(x, digits = digits, size = size_point)
   }
