+++ R/data_plot.R
@@ -101,7 +101,8 @@ data_plot <- function(x, data = NULL, ...) {
 
 #' Complete figure with its attributes
 #'
-#' The [data_plot] function usually stores information (such as title, axes labels etc.) as attributes. This function adds those information to the plot.
+#' The [data_plot] function usually stores information (such as title, axes
+#' labels etc.) as attributes. This function adds those information to the plot.
 #'
 #' @inheritParams data_plot
 #' @examples
