+++ R/plot.performance_pp_check.R
@@ -39,7 +39,7 @@ data_plot.performance_pp_check <- function(x, ...) {
 #'
 #' The `plot()` method for the `performance::pp_check()` function.
 #'
-#' @param line_alpha Alpha value of lines indicating `yrep`.
+#' @param line_alpha Numeric value specifying alpha of lines indicating `yrep`.
 #' @inheritParams data_plot
 #' @inheritParams plot.see_check_normality
 #' @inheritParams plot.see_parameters_distribution
