+++ R/plot.parameters_model.R
@@ -2,10 +2,11 @@
 #'
 #' The `plot()` method for the `parameters::model_parameters()` function.
 #'
-#' @param type Indicating the type of plot. Only applies for model parameters
-#'   from meta-analysis objects (e.g. \pkg{metafor}).
-#' @param component Indicate which component of the model should be plotted.
-#' @param weight_points Logical, if `TRUE`, for meta-analysis objects, point
+#' @param type Character indicating the type of plot. Only applies for model
+#'   parameters from meta-analysis objects (e.g. \pkg{metafor}).
+#' @param component Character indicating which component of the model should be
+#'   plotted.
+#' @param weight_points Logical. If `TRUE`, for meta-analysis objects, point
 #'   size will be adjusted according to the study-weights.
 #' @inheritParams data_plot
 #' @inheritParams plot.see_bayesfactor_parameters
