### 2021-07-22 easystats/see#178 @IndrajeetPatil
All of your concerns have now been addressed, except:
> Add links to referenced functions from other packages

We prefer not to do this because the `pkgdown` website for our package already does this.

We are not the only ones to adopt this practice. For example, if you look at `dplyr::select` roxygen [docs](https://github.com/tidyverse/dplyr/blob/ca55fce9226ce9e79f496a3e6634381311f05f47/R/select.R#L47-L49), they are not linking to external packages:

But the `pkgdown` webpage for this function [links](https://dplyr.tidyverse.org/reference/select.html#examples) it successfully:

So, it is unnecessary to replace every instance of `\code{performance::check_homogeneity()}` with `\code{\link[performance]{performance::check_homogeneity}}`.
