### 2021-07-23 #3393 @strengejacke
@jakobbossek Thanks for your comments! Let me briefly summarize what we have done to address your issues.

_Furthermore, I believe that the API/function documentation can be improved (see easystats/see#178)_

We have revised the documentation and replaced plain URLs by clickable hyperlinks. Furthermore, we explained the required data types for arguments, where missing. Finally, we improved the description of arguments to improve clarity.

_in particular argument checks are missing throughout the package (see easystats/see#175)_

We have added input checks at places where it made sense to us. However, thought debatable, we were hesitant to check input for functions that serve as extension to the ggplot2-framework, to still be consistent with that framework. The broader explanation that you were accepting can be found [here](https://github.com/easystats/see/issues/175).

_My biggest concern at the moment is that the paper is almost identical to the README file (some parts are even copy & paste). IMHO that does not make a good impression! The authors should also check whether it is really necessary to place that many code examples and the produced figures in the paper. Again, this is done in the package vignettes and the long list of examples in the README. Instead I suggest to make the paper less crowded, but more focused. However, happy to discuss the latter point_

To cover the complete easystats-suite in our paper, we believe it is good to show examples for each package. However, to avoid redundancies, we have revised the readme and the paper, and reduced the amount of examples in the paper. The paper is now shorter, yet covering the complete set of packages.

We hope that we have addressed all raised issues to your satisfaction, too. :-)
