### 2021-07-23 fce9374
turn on vignette examples
