+++ paper/paper.bib
@@ -140,7 +140,7 @@
 
 @Manual{Horikoshi2018,
     author = {Masaaki Horikoshi and Yuan Tang},
-    title = {ggfortify: Data Visualization Tools for Statistical Analysis Results},
+    title = {{ggfortify}: Data Visualization Tools for Statistical Analysis Results},
     year = {2018},
     url = {https://CRAN.R-project.org/package=ggfortify},
   }
