+++ paper.bib
@@ -1,139 +1,0 @@
-@Article{Ben-Shachar2020,
-    title = {{e}ffectsize: Estimation of Effect Size Indices and Standardized Parameters},
-    author = {Mattan S. Ben-Shachar and Daniel Lüdecke and Dominique Makowski},
-    year = {2020},
-    journal = {Journal of Open Source Software},
-    volume = {5},
-    number = {56},
-    pages = {2815},
-    publisher = {The Open Journal},
-    doi = {10.21105/joss.02815},
-    url = {https://doi.org/10.21105/joss.02815},
-  }
-
-
-@Article{Lüdecke2020parameters,
-    title = {Extracting, Computing and Exploring the Parameters of Statistical Models using {R}.},
-    volume = {5},
-    doi = {10.21105/joss.02445},
-    number = {53},
-    journal = {Journal of Open Source Software},
-    author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Dominique Makowski},
-    year = {2020},
-    pages = {2445},
-  }
-
-@Article{Lüdecke2020performance,
-    title = {{performance}: An {R} Package for Assessment, Comparison and Testing of Statistical Models},
-    author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
-    year = {2021},
-    journal = {Journal of Open Source Software},
-    volume = {6},
-    number = {60},
-    pages = {3139},
-    doi = {10.21105/joss.03139},
-  }
-
-
-@Article{Lüdecke2019,
-    title = {{insight}: A Unified Interface to Access Information from Model Objects in {R}.},
-    volume = {4},
-    doi = {10.21105/joss.01412},
-    number = {38},
-    journal = {Journal of Open Source Software},
-    author = {Daniel Lüdecke and Philip Waggoner and Dominique Makowski},
-    year = {2019},
-    pages = {1412},
-  }
-
-
-@Article{Makowski2020,
-    title = {Methods and Algorithms for Correlation Analysis in {R}.},
-    author = {Dominique Makowski and Mattan S. Ben-Shachar and Indrajeet Patil and Daniel Lüdecke},
-    doi = {10.21105/joss.02306},
-    year = {2020},
-    journal = {Journal of Open Source Software},
-    number = {51},
-    volume = {5},
-    pages = {2306},
-    url = {https://joss.theoj.org/papers/10.21105/joss.02306},
-  }
-
-@Article{Patil2021,
-    doi = {10.21105/joss.03167},
-    url = {https://doi.org/10.21105/joss.03167},
-    year = {2021},
-    publisher = {{The Open Journal}},
-    volume = {6},
-    number = {61},
-    pages = {3167},
-    author = {Indrajeet Patil},
-    title = {{Visualizations with statistical details: The {'ggstatsplot'} approach}},
-    journal = {{Journal of Open Source Software}},
-  }
-
-@Manual{Waggoner2020,
-    title = {{plotmm}: Tidy Tools for Visualizing Mixture Models},
-    author = {Philip D. Waggoner},
-    year = {2020},
-    note = {R package version 0.1.0},
-    url = {https://CRAN.R-project.org/package=plotmm},
-  }
-
-@Article{Makowski2019,
-    title = {{bayestestR}: Describing Effects and their Uncertainty, Existence and Significance within the {B}ayesian Framework.},
-    author = {Dominique Makowski and Mattan S. Ben-Shachar and Daniel Lüdecke},
-    journal = {Journal of Open Source Software},
-    doi = {10.21105/joss.01541},
-    year = {2019},
-    number = {40},
[…]
