### 2021-07-25 94a0025
Add `show_labels` argument to plots
It is more intuitive to toggle text on with `show_labels` than by setting a non-missing value for `size_text`.

Also gives the option to disable some text that was always on.
