+++ R/plot.check_outliers.R
@@ -3,6 +3,7 @@
 #' The `plot()` method for the `performance::check_outliers()`
 #' function.
 #'
+#' @param show_labels Logical. If `TRUE`, text labels are displayed.
 #' @param size_text Numeric value specifying size of text labels.
 #' @param rescale_distance Logical. If `TRUE`, distance values are rescaled
 #'   to a range from 0 to 1. This is mainly due to better catch the differences
@@ -30,6 +31,7 @@ plot.see_check_outliers <- function(x,
                                     colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                     rescale_distance = TRUE,
                                     type = c("dots", "bars"),
+                                    show_labels = TRUE,
                                     ...) {
   type <- match.arg(type)
   influential_obs <- attributes(x)$influential_obs
@@ -38,6 +40,7 @@ plot.see_check_outliers <- function(x,
   if (type == "dots" && !is.null(influential_obs) && (is.null(methods) || length(methods) == 1)) {
     .plot_diag_outliers_new(
       influential_obs,
+      show_labels = show_labels,
       size_text = size_text,
       size_line = size_line,
       dot_alpha_level = dot_alpha,
@@ -45,7 +48,7 @@ plot.see_check_outliers <- function(x,
     )
   } else {
     if (length(methods == 1)) {
-      .plot_diag_outliers(x, size_text, rescale_distance)
+      .plot_diag_outliers(x, show_labels = show_labels, size_text = size_text, rescale_distance = rescale_distance)
     } else {
       .plot_outliers_multimethod(x, rescale_distance)
     }
@@ -81,7 +84,7 @@ data_plot.check_outliers <- function(x, data = NULL, rescale_distance = TRUE, ..
 }
 
 
-.plot_diag_outliers <- function(x, size_text = 3.5, rescale_distance = TRUE) {
+.plot_diag_outliers <- function(x, show_labels = TRUE, size_text = 3.5, rescale_distance = TRUE) {
   d <- data_plot(x, rescale_distance = rescale_distance)
   d$Id <- 1:nrow(d)
   d$Outliers <- as.factor(attr(x, "data", exact = TRUE)[["Outlier"]])
@@ -106,7 +109,7 @@ data_plot.check_outliers <- function(x, data = NULL, rescale_distance = TRUE, ..
     x_lab <- method
   }
 
-  if (is.null(size_text)) size_text <- 3.5
+  size_text <- size_text %||% 3.5
 
   p <- ggplot(d, aes(x = .data$Distance, fill = .data$Outliers, label = .data$Id)) +
     geom_histogram() +
@@ -130,10 +133,12 @@ data_plot.check_outliers <- function(x, data = NULL, rescale_distance = TRUE, ..
   }
 
 
-  if (requireNamespace("ggrepel", quietly = TRUE)) {
-    p <- p + ggrepel::geom_text_repel(y = 2.5, size = size_text, na.rm = TRUE)
-  } else {
-    p <- p + geom_text(y = 2.5, size = size_text, na.rm = TRUE)
+  if (isTRUE(show_labels)) {
+    if (requireNamespace("ggrepel", quietly = TRUE)) {
+      p <- p + ggrepel::geom_text_repel(y = 2.5, size = size_text, na.rm = TRUE)
+    } else {
+      p <- p + geom_text(y = 2.5, size = size_text, na.rm = TRUE)
+    }
   }
 
   p
