+++ R/plot.parameters_pca.R
@@ -70,6 +70,7 @@ plot.see_parameters_pca <- function(x,
                                     size_text = 3.5,
                                     text_color = "black",
                                     size = 1,
+                                    show_labels = TRUE,
                                     ...) {
   type <- match.arg(type)
 
@@ -106,13 +107,17 @@ plot.see_parameters_pca <- function(x,
       scale_color_gradientn(colours = c("#cd201f", "#ffffff", "#0077B5"), limits = c(-1, 1))
   }
 
+  if (isTRUE(show_labels)) {
+    p <- p +
+      geom_text(
+        aes(x = abs(.data$y), label = round(.data$y, 2)),
+        color = text_color,
+        size = size_text,
+        nudge_y = .15
+      )
+  }
+
   p <- p +
-    geom_text(
-      aes(x = abs(.data$y), label = round(.data$y, 2)),
-      color = text_color,
-      size = size_text,
-      nudge_y = .15
-    ) +
     guides(fill = "none", color = "none") +
     scale_x_continuous(
       limits = c(0, 1.25),
