### 2021-07-25 034df73
easycormatrix: Change `show_values` to `show_labels` for consistency
Accepts `show_labels` as an alias for backwards compatibility. Also fix bug where `show_p` was ignored if `show_values = FALSE`.
