+++ R/plot.check_outliers.R
@@ -83,7 +83,7 @@ data_plot.check_outliers <- function(x, data = NULL, rescale_distance = TRUE, ..
   d_long
 }
 
-
+#' @importFrom rlang `%||%`
 .plot_diag_outliers <- function(x, show_labels = TRUE, size_text = 3.5, rescale_distance = TRUE) {
   d <- data_plot(x, rescale_distance = rescale_distance)
   d$Id <- 1:nrow(d)
