+++ R/utils.R
@@ -185,7 +185,3 @@
 .is_integer <- function(x) {
   is.numeric(x) && all(floor(x) == x, na.rm = T)
 }
-
-`%||%` <- function (x, y) {
-  if (is.null(x)) y else x
-}
