+++ NAMESPACE
@@ -176,6 +176,7 @@ export(theme_radar)
 export(theme_radar_dark)
 import(ggplot2)
 importFrom(rlang,.data)
+importFrom(rlang,`%||%`)
 importFrom(stats,density)
 importFrom(stats,na.omit)
 importFrom(stats,xtabs)
