### 2021-07-29 #3393 @fboehm
One minor fix is needed:

line 34: "and the *plotmm* package". Need to add the word "and".
