### 2021-07-29 #3393 @fboehm
I confirmed that the dois resolve to the cited works.
