### 2021-07-29 #3393 @strengejacke
@fboehm Thanks! I have created a new release from the latest version, including the final fixes you suggested for the paper:
https://github.com/easystats/see/releases/tag/0.6.5

This release is also archived on zenodo, DOI: 10.5281/zenodo.5146411
