### 2022-01-24 #3393 @dr-suhadalishaheed
> @MatthewSmith430, @jakobbossek – many thanks for your reviews here and to @fboehm for editing this submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
>
>
>
> @strengejacke – your paper is now accepted and published in JOSS :zap::rocket::boom:

- [ ] @arfon
